docs(README): Create overview and error examples
Describe the purpose of this repository and why the code is
intentionally broken. Espcially important is that the error
can not be fixed at the level of this repository as the
package from `demo-main`, even though accessible, can not be resolved
and fetched by the go tool to compile the project.